+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+## (0.3.9) 21/10/24
+### Changed
+- cambio en el codigo de chess.py, nuevos metodos agregados para manejar el juego
+### Issues
+- arrgelar problemas al realizar movimientos dentro del tablero en chess.py
 ## (0.3.8) 21/10/24
 ### Added
 - nuevos metodos para board.py